+++ blas/cpu/NativeOps.cpp
@@ -29,7 +29,7 @@
 #include <unistd.h>
 
 #include <layers/layers_factory.h>
-#include <ops/declarable/declarable_ops.h>
+#include <ops/declarable/DeclarableOp.h>
 #include <ops/declarable/generic/parity_ops.h>
 #include <ops/declarable/generic/third_party.h>
