+++ include/ops/declarable/DeclarableOp.h
@@ -123,98 +123,7 @@ namespace nd4j {
             Nd4jStatus validateArguments(Block<T>& block);
         };
 
-
-        template <typename T>
-        class DeclarableReductionOp : public nd4j::ops::DeclarableOp<T> {
-        protected:
-            /**
-             * This method executes this Op
-             */
-            virtual Nd4jStatus validateAndExecute(Block<T>& block) = 0;
-        public:
-            DeclarableReductionOp(int numInputs, int numOutputs, const char *opName, bool allowsInplace, int tArgs, int iArgs) : nd4j::ops::DeclarableOp<T>(numInputs, numOutputs, opName, allowsInplace, tArgs, iArgs) {
-                //
-            }
-
-            ~DeclarableReductionOp()  {
-                //
-            }
-
-            ShapeList* calculateOutputShape(ShapeList* inputShape, nd4j::graph::Block<T>& block);
-        };
-
-        template <typename T>
-        class DeclarableCustomOp : public nd4j::ops::DeclarableOp<T> {
-        protected:
-            /**
-             * This method executes this Op
-             */
-            virtual Nd4jStatus validateAndExecute(Block<T>& block) = 0;
-        public:
-            DeclarableCustomOp(int numInputs, int numOutputs, const char *opName, bool allowsInplace, int tArgs, int iArgs) : nd4j::ops::DeclarableOp<T>(numInputs, numOutputs, opName, allowsInplace, tArgs, iArgs) {
-                //
-            }
-
-            ~DeclarableCustomOp()  {
-                //
-            }
-
-            virtual ShapeList* calculateOutputShape(ShapeList* inputShapes, nd4j::graph::Block<T>& block) = 0;
-        };
-
     }
 }
-
-
-
-template <typename T>
-nd4j::ShapeList* nd4j::ops::DeclarableReductionOp<T>::calculateOutputShape(nd4j::ShapeList* inputShape, nd4j::graph::Block<T>& block)  {
-    int numDims = block.getIArguments()->at(0);
-    std::vector<int> dims;
-    for (int e = 0; e < numDims; e++)
-        dims.push_back(block.getIArguments()->at(e+1));
-
-    if (numDims > 1)
-        std::sort(dims.begin(), dims.end());
-
-    // special case - output is scalar
-    if (numDims == 1 && dims.at(0) == MAX_INT) {
-        int* newShape;
-        ALLOCATE(newShape, block.getWorkspace(), 8, int);
-
-        newShape[0] = 2;
-        newShape[1] = 1;
-        newShape[2] = 1;
-        newShape[3] = 1;
-        newShape[4] = 1;
-        newShape[5] = 0;
-        newShape[6] = 1;
-        newShape[7] = 99;
-
-        return new ShapeList(newShape);
-    }
-
-    shape::TAD tad(inputShape->at(0), dims.data(), numDims);
-    tad.createTadOnlyShapeInfo();
-
-    Nd4jIndex tadLength = shape::tadLength(inputShape->at(0), dims.data(), numDims);
-    Nd4jIndex numTads = shape::length(inputShape->at(0)) /  tadLength;
-
-    int* newShape;
-    ALLOCATE(newShape, block.getWorkspace(), shape::shapeInfoLength(2), int);
-
-    newShape[0] = 2;
-    newShape[1] = 1;
-    newShape[2] = numTads;
-    newShape[3] = numTads;
-    newShape[4] = 1;
-    newShape[5] = 0;
-    newShape[6] = 1;
-    newShape[7] = 99;
-
-    return new ShapeList(newShape);
-}
-
-
 
 #endif //LIBND4J_DECLARABLE_OPS_H
