+++ include/ops/declarable/generic/third_party.h
@@ -13,7 +13,7 @@
 #include <ops/ops.h>
 #include <loops/random.h>
 #include <NDArray.h>
-#include <ops/declarable/declarable_ops.h>
+#include <ops/declarable/DeclarableOp.h>
 #include <ops/declarable/OpRegistrator.h>
 #include <NDArrayFactory.h>
