+++ tests/layers_tests/CyclicTests.h
@@ -11,7 +11,7 @@
 #include <Node.h>
 #include <graph/Variable.h>
 #include <graph/VariableSpace.h>
-#include <ops/declarable/declarable_ops.h>
+#include <ops/declarable/DeclarableOp.h>
 #include <ops/declarable/generic/convo/convo_ops.h>
 
 using namespace nd4j::graph;
